+++ kea/hdl/axi/axi_stream_connector/_axis_connector.py
@@ -1,7 +1,7 @@
 from myhdl import block
 
 from kea.hdl.axi import AxiStreamInterface
-from kea.hdl.signal_handling import signal_assigner
+from kea.hdl.signal_handling import sig_assigner
 
 @block
 def axis_connector(clock, axis_source, axis_sink):
@@ -28,34 +28,34 @@ def axis_connector(clock, axis_source, axis_sink):
 
     # Connect the axis source signals to the axis sink signals
     return_objects.append(
-        signal_assigner(axis_source.TVALID, axis_sink.TVALID))
+        sig_assigner(axis_source.TVALID, axis_sink.TVALID))
     return_objects.append(
-        signal_assigner(axis_source.TDATA, axis_sink.TDATA))
+        sig_assigner(axis_source.TDATA, axis_sink.TDATA))
     return_objects.append(
-        signal_assigner(axis_sink.TREADY, axis_source.TREADY))
+        sig_assigner(axis_sink.TREADY, axis_source.TREADY))
 
     if axis_source.TDEST_width is not None:
         return_objects.append(
-            signal_assigner(axis_source.TDEST, axis_sink.TDEST))
+            sig_assigner(axis_source.TDEST, axis_sink.TDEST))
 
     if axis_source.TID_width is not None:
         return_objects.append(
-            signal_assigner(axis_source.TID, axis_sink.TID))
+            sig_assigner(axis_source.TID, axis_sink.TID))
 
     if axis_source.TUSER_width is not None:
         return_objects.append(
-            signal_assigner(axis_source.TUSER, axis_sink.TUSER))
+            sig_assigner(axis_source.TUSER, axis_sink.TUSER))
 
     if hasattr(axis_source, 'TLAST'):
         return_objects.append(
-            signal_assigner(axis_source.TLAST, axis_sink.TLAST))
+            sig_assigner(axis_source.TLAST, axis_sink.TLAST))
 
     if hasattr(axis_source, 'TSTRB'):
         return_objects.append(
-            signal_assigner(axis_source.TSTRB, axis_sink.TSTRB))
+            sig_assigner(axis_source.TSTRB, axis_sink.TSTRB))
 
     if hasattr(axis_source, 'TKEEP'):
         return_objects.append(
-            signal_assigner(axis_source.TKEEP, axis_sink.TKEEP))
+            sig_assigner(axis_source.TKEEP, axis_sink.TKEEP))
 
     return return_objects